requirements.txt
Django==1.7.1
psycopg2>=2.5.2
Pillow==2.7.0
PyYAML==3.11
django-admin-bootstrapped==2.3.1
django-appconf==0.6
django-filter==0.9.2
django-filters==0.1.0
django-imagekit==3.2.5
djangorestframework>=2.4,<3.0
factory-boy==2.4.1
pilkit==1.1.12
six==1.9.0
wsgiref==0.1.2
